if a populate event is fired on a biome world regenerate the chunk.
This is needed due to loading a chunk not causing population. Population only occurs when enough chunks are loaded in an area, so we load the chunks in the area and wait to see if a populate event is fired. This causes it to go much slower but actually works ¯\_(ツ)_/¯